feat(backend): impelement sign in and everything childy for it

+++ src/Reminder.Backend/Reminder/Reminder.Application/Providers/JwtProvider.cs
@@ -19,7 +19,7 @@ public class JwtProvider
         _refreshSessionConfiguration = refreshSessionConfiguration;
     }
 
-    public string GenerateUserJwt(int userId, JwtType jwtType)
+    public string GenerateUserJwt(long userId, JwtType jwtType)
     {
         var claims = new List<Claim>
         {
